sebességváltozás bugfix maxspeed eltávolítva

+++ src/Model/Bullet.java
@@ -8,7 +8,7 @@ public class Bullet extends GameElements {
     private double angle;
     private double speed;
 
-    public Bullet(Point location, double angle, int speed) {
+    public Bullet(Point location, double angle, double speed) {
         super(location,5);
         this.location=new Point(
                 (int) ((Resources.PlayerRobot1Image.getWidth()/2) * Math.cos(Math.toRadians(angle))) + (int) location.getX(),
